DRY mocks in tests
