+++ frontend/javascripts/test/mocks/globals.mock.ts
@@ -1,0 +1,12 @@
+import mock from "mock-require";
+
+// @ts-ignore
+global.performance = {
+  now: () => Date.now(),
+};
+mock("libs/window", {
+  requestAnimationFrame: () => {},
+  document: {
+    getElementById: () => null,
+  },
+});
