+++ frontend/javascripts/test/model/cuckoo_table.spec.ts
@@ -1,66 +1,10 @@
-import * as THREE from "three";
 import mock from "mock-require";
 import test, { ExecutionContext } from "ava";
 import _ from "lodash";
 import { Vector3 } from "oxalis/constants";
-/*
- * Note that RGB textures are currently not tested in this spec.
- * If tests were added, the following Map would not be sufficient, anymore,
- * since RGBAFormat is also used for 3 channels which would make the key not unique.
- */
-const formatToChannelCount = new Map([
-  [THREE.RedFormat, 1],
-  [THREE.RedIntegerFormat, 1],
-  [THREE.RGFormat, 2],
-  [THREE.RGIntegerFormat, 2],
-  [THREE.RGBAFormat, 4],
-  [THREE.RGBAIntegerFormat, 4],
-]);
 
-// @ts-ignore
-global.performance = {
-  now: () => Date.now(),
-};
-mock("libs/window", {
-  requestAnimationFrame: () => {},
-  document: {
-    getElementById: () => null,
-  },
-});
-mock(
-  "libs/UpdatableTexture",
-  class UpdatableTexture {
-    texture: Uint8Array = new Uint8Array();
-    width: number = 0;
-    height: number = 0;
-    channelCount: number;
-
-    constructor(_width: number, _height: number, format: any) {
-      this.channelCount = formatToChannelCount.get(format) || 0;
-      if (this.channelCount === 0) {
-        throw new Error("Format could not be converted to channel count");
-      }
-    }
-
-    update(src: Float32Array | Uint8Array, x: number, y: number, _width: number, _height: number) {
-      this.texture.set(src, y * this.width + x);
-    }
-
-    setRenderer() {}
-
-    setSize(width: number, height: number) {
-      this.texture = new Uint8Array(width * height * this.channelCount);
-      this.width = width;
-      this.height = height;
-    }
-
-    isInitialized() {
-      return true;
-    }
-  },
-);
-
-// import { CuckooTable } from "oxalis/model/bucket_data_handling/cuckoo_table";
+import "test/mocks/globals.mock";
+import "test/mocks/updatable_texture.mock";
 
 const { CuckooTable } = mock.reRequire("oxalis/model/bucket_data_handling/cuckoo_table");
